small refacto on virtual activities
